requirements_redaction.txt
# Function dependencies, for example:
# package>=version
bcrypt==3.2.0
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cloudstorage==0.11.0
cryptography==36.0.2
ffmpeg==1.4
ffmpeg-python==0.2.0
future==0.18.2
gcloud==0.18.3
google-api-core==2.7.2
google-auth==2.6.6
google-cloud==0.34.0
google-cloud-core==2.3.0
google-cloud-datastore==2.5.1
google-cloud-dlp==3.7.0
google-cloud-pubsub==2.12.0
google-cloud-secret-manager==2.10.0
google-cloud-speech==2.13.1
google-cloud-storage==2.3.0
google-crc32c==1.3.0
google-resumable-media==2.3.2
googleapis-common-protos==1.56.0
grpc-google-iam-v1==0.12.4
grpcio==1.44.0
grpcio-status==1.44.0
httplib2==0.20.4
idna==3.3
inflection==0.5.1
numpy==1.22.3
oauth2client==4.1.3
pandas==1.4.2
paramiko==2.10.3
proto-plus==1.20.3
protobuf==3.20.1
pubsub==0.1.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydub==0.25.1
PyNaCl==1.5.0
pyparsing==3.0.8
pysftp==0.2.9
python-dateutil==2.8.2
python-magic==0.4.25
pytz==2022.1
requests==2.27.1
rsa==4.8
six==1.16.0
typing==3.7.4.3
urllib3==1.26.9
Wave==0.0.2
Werkzeug==2.1.2
